requirements.txt
boto3==1.13.1
botocore==1.16.1
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-MySQLdb==0.2.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.1.2
python-dateutil==2.8.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.9
Werkzeug==1.0.1
xlrd==1.2.0

